code migration to update Details controller and username, password added in prod properties file
