+++ demo/src/main/resources/application-prod.properties
@@ -3,8 +3,8 @@ server.port=8281
 
 #MYSQL
 spring.datasource.url=jdbc:mysql://localhost:3306/demo
-spring.datasource.username=root
-spring.datasource.password=root
+spring.datasource.username=database-1.ct0msymoo2bv.ap-south-1.rds.amazonaws.com
+spring.datasource.password=insdb123
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 #Hibernate
